It's sad to have the IDE dictate code:
VSCode's test runner & test explorer explodes with
the huge amounts of test data combinations we run with.
The only solution I could find was to remove test.each
calls and instead put the loops in it calls.
The drawback of-course is when a test explodes,
you don't get the full list of test data combinations